refactor: break up conditions service
