+++ refiner/app/api/v1/configurations/testing.py
@@ -23,7 +23,7 @@ from app.db.demo.model import Condition
 from app.db.pool import AsyncDatabaseConnection, get_db
 from app.db.users.model import DbUser
 from app.services.aws.s3 import upload_refined_file_package
-from app.services.conditions import filter_refined_files_by_diff_rendering
+from app.services.conditions.refinement import filter_refined_files_by_diff_rendering
 from app.services.file_io import (
     ZipFileItem,
     ZipFilePackage,
