+++ refiner/app/services/aws/s3_keys.py
@@ -1,25 +1,6 @@
-from urllib.parse import urlparse
-from uuid import UUID
+from app.services.conditions.parsing import extract_uuid_from_canonical_url
 
 S3_CONFIGURATION_DIR_PREFIX = "configurations"
-
-
-def _extract_uuid_from_canonical_url(url: str) -> UUID:
-    """
-    Given a `canonical_url`, extracts and returns the UUID at the end of it.
-
-    For example, given `https://tes.tools.aimsplatform.org/api/fhir/ValueSet/c435a017-41e6-4030-b6d1-0bda2eb05b1f`
-    the function returns `c435a017-41e6-4030-b6d1-0bda2eb05b1f`
-
-    Args:
-        url (str): The canonical URL
-
-    Returns:
-        UUID: The UUID at the end of the canonical URL
-    """
-    parsed = urlparse(url)
-    last_segment = parsed.path.rstrip("/").split("/")[-1]
-    return UUID(last_segment)
 
 
 def get_jurisdiction_directory(jurisdiction_id: str) -> str:
@@ -46,7 +27,7 @@ def get_parent_directory_key(jurisdiction_id: str, canonical_url: str) -> str:
     Returns:
         str: Full S3 key to the activation file parent directory
     """
-    uuid = _extract_uuid_from_canonical_url(url=canonical_url)
+    uuid = extract_uuid_from_canonical_url(canonical_url=canonical_url)
     return f"{get_jurisdiction_directory(jurisdiction_id=jurisdiction_id)}/{uuid}"
